fix: Fix `test_tcp_server_started` to close the connection (#363)
Fix `test_tcp_server_started` to use a context manager, in order to
close the connection to the `TcpFakeServer` when done.  Otherwise,
the test relies on GC closing the connection in order for the server
thread to finish -- which does not happen with PyPy, and causes `pytest`
to hang after running the test suite.